+++ tests/test_checkers/test_noqa.py
@@ -26,4 +26,5 @@ def test_noqa_fixture(absolute_path):
     )
     stdout, _ = process.communicate()
 
+    print(stdout)
     assert stdout.count(b'Z') == 0