+++ wemake_python_styleguide/visitors/wrong_import.py
@@ -2,7 +2,9 @@
 
 import ast
 
+from wemake_python_styleguide.constants import FUTURE_IMPORTS_WHITELIST
 from wemake_python_styleguide.errors import (
+    FutureImportViolation,
     LocalFolderImportViolation,
     NestedImportViolation,
 )
@@ -21,24 +23,33 @@ class WrongImportVisitor(BaseNodeVisitor):
             return node.names[0].name
         return '.'
 
-    def _check_nested_import(self, node: ast.AST, text: str):
+    def _check_nested_import(self, node: ast.AST):
+        text = self._get_error_text(node)
         parent = getattr(node, 'parent', None)
         if not isinstance(parent, ast.Module):
             self.add_error(NestedImportViolation(node, text=text))
 
-    def _check_local_import(self, node: ast.ImportFrom, text: str):
+    def _check_local_import(self, node: ast.ImportFrom):
+        text = self._get_error_text(node)
         if node.level != 0:
             self.add_error(LocalFolderImportViolation(node, text=text))
 
+    def _check_future_import(self, node: ast.ImportFrom):
+        if node.module == '__future__':
+            for alias in node.names:
+                if alias.name not in FUTURE_IMPORTS_WHITELIST:
+                    self.add_error(
+                        FutureImportViolation(node, text=alias.name),
+                    )
+
     def visit_Import(self, node: ast.Import):
         """Used to find nested `import` statements."""
-        text = self._get_error_text(node)
-        self._check_nested_import(node, text)
+        self._check_nested_import(node)
         self.generic_visit(node)
 
     def visit_ImportFrom(self, node: ast.ImportFrom):
         """Used to find nested `from import` statements and local imports."""
-        text = self._get_error_text(node)
-        self._check_local_import(node, text)
-        self._check_nested_import(node, text)
+        self._check_local_import(node)
+        self._check_nested_import(node)
+        self._check_future_import(node)
         self.generic_visit(node)